Added Behavior Graph Visualization
A new method in SimulationData now uses the Networkx module to output a
graph of the champion bot's behavior along with the other metrics.
The simulation no longer saves the champions behavior as text, only a
graph.
Running the simulation may report in an 'alloc: invalid block' error
after the end of the simulation when Python shuts down. It does not
affect the simulation or its output and seems to arise from Networkx or
matplotlib.