Lots of Bug Fixes
Bot age also displays as a percent of maximum age.
Fixed a bug caused by forgetting to add a behavior function to a list
that which results in crashing during graphing of champ intelligence.
'Push away' signals now work correctly. Also made them easier to spot.
Fixed a mistake in the image font.